Refactorizar uso correcto de MVVM
