+++ tech_nebrios_tracker/lib/data/models/constantes.dart
@@ -1,0 +1,9 @@
+class APIRutas{
+  static const String BASEURL = "http://localhost:3000";
+
+  static const String CHAROLA = "$BASEURL/charola";
+
+  static const String USUARIO = "$BASEURL/usuario";
+
+  static const String ALIMENTACION = "$BASEURL/alimentacion";
+}
