+++ tech_nebrios_tracker/test/mocks/mocks.dart
@@ -1,9 +1,9 @@
 import 'package:mockito/annotations.dart';
 import 'package:tech_nebrios_tracker/domain/alimentacion_domain.dart';
-import 'package:tech_nebrios_tracker/data/repositories/alimento_repository.dart';
+import 'package:tech_nebrios_tracker/data/repositories/alimentacion_repository.dart';
 
 @GenerateMocks([
   EditarAlimentoCasoUso, 
-  AlimentoRepository     
+  AlimentacionRepository     
 ])
 void main() {}
